restructure k4simdelphes as its own repo
