+++ K4SIMDELPHESVersion.h
@@ -1,0 +1,4 @@
+#ifndef K4SIMDELPHESVersion_h
+#define K4SIMDELPHESVersion_h
+#define K4SIMDELPHES_VERSION @K4SIMDELPHES_VERSION@
+#endif
