+++ cmake/K4SIMDELPHESConfig.cmake.in
@@ -1,0 +1,21 @@
+# - Config file for the K4SIMDELPHES package
+
+# - Define exported version
+set(K4SIMDELPHES_VERSION "@PROJECT_VERSION@")
+
+# - Init CMakePackageConfigHelpers
+@PACKAGE_INIT@
+
+# - Create relocatable paths to headers.
+# NOTE: Do not strictly need paths as all usage requirements are encoded in
+# the imported targets created later.
+set_and_check(K4SIMDELPHES_INCLUDE_DIR "@PACKAGE_CMAKE_INSTALL_INCLUDEDIR@")
+
+include(CMakeFindDependencyMacro)
+find_dependency(Delphes REQUIRED)
+find_dependency(EDM4HEP REQUIRED)
+
+# - Include the targets file to create the imported targets that a client can
+# link to (libraries) or execute (programs)
+include("${CMAKE_CURRENT_LIST_DIR}/K4SIMDELPHESTargets.cmake")
+
